refactors on cli structure
