+++ packages/cli/src/utils/formatting.js
